chore(optimizations): migration safe-point 3 (clippy clean)

+++ crates/varn-debug/src/bytecode.rs
@@ -532,7 +532,7 @@ fn print_proto(proto: &FunctionProto, depth: usize, total: &mut usize) {
                 // Los descriptores de captura nombran SLOTS del frame padre
                 // (`is_local=1`) o upvalues heredadas; sin verlos es imposible
                 // auditar la interacción entre captura y register allocation.
-                let mut uvs = Vec::with_capacity(uv_count as usize);
+                let mut uvs = Vec::with_capacity(uv_count);
                 for _ in 0..uv_count {
                     let d = w!();
                     uvs.push(if hi(d) != 0 {
